fix!: reorganize chart and vegalite and rename vegalite component
BREAKING CHANGE: rename vagaLiteChart to vegaLite
