--- src/components/chart/vegaLiteChart/vegaLiteChart.css
@@ -1,8 +1,0 @@
-.rustic-chart {
-  width: 100%;
-  height: 100%;
-}
-
-.rustic-vega-lite-chart {
-  flex: 1;
-}
