+++ src/components/visualization/vegaLite/vegaLite.tsx
@@ -1,4 +1,4 @@
-import './vegaLiteChart.css'
+import './vegaLite.css'
 
 import Typography from '@mui/material/Typography'
 import Stack from '@mui/system/Stack'
@@ -6,10 +6,10 @@ import useTheme from '@mui/system/useTheme'
 import React, { useEffect, useRef, useState } from 'react'
 import { default as VegaEmbed } from 'vega-embed'
 
-import type { VegaLiteChartData } from '../../types'
+import type { VegaLiteData } from '../../types'
 
-/** The `VegaLiteChart` component is a versatile tool for visualizing data using the Vega-Lite grammar. With support for various chart types, it empowers users to create engaging and informative data visualizations effortlessly. */
-function VegaLiteChart(props: VegaLiteChartData) {
+/** The `VegaLite` component is a versatile tool for visualizing data using the Vega-Lite grammar. With support for various graphic types, it empowers users to create engaging and informative data visualizations effortlessly. */
+function VegaLite(props: VegaLiteData) {
   const chartRef = useRef<HTMLDivElement>(null)
   const [hasError, setHasError] = useState<boolean>(false)
   const isDarkTheme = useTheme().palette.mode === 'dark'
@@ -49,26 +49,22 @@ function VegaLiteChart(props: VegaLiteChartData) {
     return <Typography variant="body2">Failed to load the chart.</Typography>
   } else {
     return (
-      <Stack direction="column" className="rustic-chart">
+      <Stack direction="column" className="rustic-vega-lite-container">
         {props.title && (
           <Typography variant="subtitle2">{props.title}</Typography>
         )}
         {props.description && (
           <Typography variant="caption">{props.description}</Typography>
         )}
-        <div
-          ref={chartRef}
-          className="rustic-vega-lite-chart"
-          data-cy="vega-lite-chart"
-        />
+        <div ref={chartRef} className="rustic-vega-lite" data-cy="vega-lite" />
       </Stack>
     )
   }
 }
 
-export default VegaLiteChart
+export default VegaLite
 
-VegaLiteChart.defaultProps = {
+VegaLite.defaultProps = {
   theme: {
     dark: 'dark' as const,
   },
